Clean-up and move asarray and argument to _funcs module
