+++ ndonnx/_elementwise.py
@@ -7,8 +7,9 @@ Each function directly dispatches to the inner typed array.
 
 import builtins
 
-from ._array import Array, asarray
-from ._dtypes import DType
+import ndonnx as ndx
+from ndonnx import Array, DType
+
 from ._typed_array import funcs as tyfuncs
 
 
@@ -91,8 +92,8 @@ def clip(
             min_max.append(max)
     dtype = tyfuncs.result_type(x.dtype, *min_max)
 
-    min_ = None if min is None else asarray(min, dtype=dtype)._tyarray
-    max_ = None if max is None else asarray(max, dtype=dtype)._tyarray
+    min_ = None if min is None else ndx.asarray(min, dtype=dtype)._tyarray
+    max_ = None if max is None else ndx.asarray(max, dtype=dtype)._tyarray
     return Array._from_tyarray(x._tyarray.clip(min=min_, max=max_))
