+++ ndonnx/_from_numpy_dtype.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo 2023-2024
+# Copyright (c) QuantCo 2023-2025
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
@@ -30,9 +30,9 @@ def from_numpy_dtype(
         unit = datetime.validate_unit(unit_str)
         return datetime.TimeDelta64DType(unit=unit)
 
-    # "T" i.e. "Text" is the kind used for `StringDType` in numpy >= 2
+    # We don't support "T" ("Text" is the kind used for `StringDType` in numpy >= 2), yet.
     # See https://numpy.org/neps/nep-0055-string_dtype.html#python-api-for-stringdtype
-    if np_dtype.kind in ["U", "T", "i", "b", "f", "u"]:
+    if np_dtype.kind in ["U", "i", "b", "f", "u"]:
         return onnx.from_numpy_dtype(np_dtype)
 
     raise ValueError(f"'{np_dtype}' does not have a corresponding ndonnx data type")
